[AN] refactor: NetworkModule로 LoggingInterceptor 로직 통합 #740
